Consumer for trends to Cassandra
This program consumes the from the Kafka queue and feeds it to a Cassandra table.
The KEYSPACE for Cassandra used is graphy. The table name is trend.
The table has trend_id and trend_name. trend_id is the PK and of type UUID. trend_name is of type text.
All dependencies for java files by me will be in a separate folder called java dependencies Shayon.